Create a yearly review page (#53)
A simple yearly review page that summarizes your workouts by frequency. Lots of potential here, just wanted to make the page exist and give basic stats!

To minimize reads and calculations, the yearly review is stored in a new collection and is calculated only one time if a user doesn't already have a yearly review entry for the existing year. A good future improvement is for the summaries to be calculated by a cloud function rather than the client.

<img width="625" alt="yearly review screenshot" src="https://user-images.githubusercontent.com/1589186/103731135-4b7ce900-4f99-11eb-8ea4-cd3ec0daa0e7.png">