add an emoji to each workout (#54)
In this update, Workouts are now each assigned an emoji to represent them on the calendar. This is so the symbolism of each work out are more apparent at the calendar view.

![create workout form with emoji selection](https://user-images.githubusercontent.com/1589186/105564269-4a0c2a00-5cd6-11eb-8889-fda1cb5714ce.png)

Because I have no idea how you're supposed to do a database migration with schemaless db's, I just picked a random emoji for each workout when you first load this page. You can pick the emoji by going into the workouts page and clicking on the emoji.

The workout emojis will show up on your calendar: 
<img width="400" alt="workout calendar light" src="https://user-images.githubusercontent.com/1589186/105569971-1a234d80-5cfb-11eb-9743-af1ac88aef5f.png">
<img width="400" alt="workout calendar dark" src="https://user-images.githubusercontent.com/1589186/105569973-1bed1100-5cfb-11eb-84da-df2dd7479c5b.png">

Also includes a feature to highlight active days, which is on by default (there's no way to turn it off currently):
<img width="400" alt="workout calendar light, with active days highlighted" src="https://user-images.githubusercontent.com/1589186/105570841-7b99eb00-5d00-11eb-84f6-fde36efb0d27.png">
<img width="400" alt="workout calendar dark, with active days highlighted" src="https://user-images.githubusercontent.com/1589186/105570844-7dfc4500-5d00-11eb-80c4-fc4d68b46806.png">


And also show up in each journal entry:
![Journal entry with workout emoji](https://user-images.githubusercontent.com/1589186/105564268-49739380-5cd6-11eb-82e6-ae143ab55d82.png)