alphabetize workouts after fetching (#55)
It bothers me that workouts aren't alphabetized in the list, it makes them hard to find. You can't use the `.orderBy()` clause of a firebase query because that is case sensitive. That means the capitalised workouts would come before the uncapitalised workouts u_u. To get around this, I just sort in the browser.

I also cleaned up the code to cancel the watchers when components unmount.